+++ packages/client/src/enums/index.ts
@@ -1,4 +1,5 @@
 export * from './auth'
 export * from './gov'
+export * from './operation'
 export * from './staking'
 export * from './tokenmanager'