+++ packages/client/src/types/identity.ts
@@ -23,18 +23,3 @@ export type IdentityNode = {
     childrenHash: string
   }
 }
-
-export type IdentityParams = {
-  params: {
-    lcgA: string
-    lcgB: string
-    lcgMod: string
-    lcgValue: string
-    identityContractAddress: string
-    chainName: string
-    GISTHash: string
-    GISTUpdatedTimestamp: string
-    treapRootKey: string
-    statesWaitingForSign: [string]
-  }
-}