+++ packages/zkp-gen-iden3/src/zkp-gen.ts
@@ -22,7 +22,12 @@ import {
   Proof,
 } from '@iden3/js-merkletree'
 import type { VerifiableCredentials } from '@rarimo/auth-zkp-iden3'
-import type { MerkleProof, OperationProof, RarimoQuerier } from '@rarimo/client'
+import type {
+  MerkleProof,
+  Operation,
+  OperationProof,
+  RarimoQuerier,
+} from '@rarimo/client'
 import { type Identity } from '@rarimo/identity-gen-iden3'
 import { isString, omit } from '@rarimo/shared'
 import {
@@ -30,8 +35,7 @@ import {
   getCoreChainStateInfo,
   getGISTProof,
   getGISTRootInfo,
-  getIdentityNode,
-  getIdentityParams,
+  getOperation,
   unmarshalBinary,
 } from '@rarimo/shared-zkp-iden3'
 import { Buffer } from 'buffer'
@@ -93,6 +97,7 @@ export class ZkpGen<T extends QueryVariableNameAbstract> {
   public targetStateDetails?: Awaited<ReturnType<typeof getGISTRootInfo>>
   public coreStateDetails?: Awaited<ReturnType<typeof getCoreChainStateInfo>>
   public operationProof?: OperationProof
+  public operation?: Operation
   public merkleProof?: MerkleProof
 
   public circuitWasm?: Uint8Array
@@ -654,16 +659,19 @@ export class ZkpGen<T extends QueryVariableNameAbstract> {
     targetStateDetails,
     coreStateDetails,
     operationProof,
+    operation,
     merkleProof,
   }: {
     targetStateDetails?: Awaited<ReturnType<typeof getGISTRootInfo>>
     coreStateDetails?: Awaited<ReturnType<typeof getCoreChainStateInfo>>
     operationProof?: OperationProof
+    operation?: Operation
     merkleProof?: MerkleProof
   }) {
     this.targetStateDetails = targetStateDetails
     this.coreStateDetails = coreStateDetails
     this.operationProof = operationProof
+    this.operation = operation
     this.merkleProof = merkleProof
   }
 
@@ -688,6 +696,10 @@ export class ZkpGen<T extends QueryVariableNameAbstract> {
     this.operationProof = await querier.getOperationProof(operationIndex)
   }
 
+  public async loadOperation(querier: RarimoQuerier, operationIndex: string) {
+    this.operation = await getOperation(querier, operationIndex)
+  }
+
   public async loadMerkleProof(querier: RarimoQuerier, issuerId: string) {
     this.merkleProof = await querier.getMerkleProof(issuerId)
   }
@@ -700,25 +712,18 @@ export class ZkpGen<T extends QueryVariableNameAbstract> {
     )
   }
 
-  public async loadParamsForTransitState(
-    querier: RarimoQuerier,
-    opts?: {
-      operationProof?: OperationProof
-    },
-  ) {
+  public async loadParamsForTransitState(opts?: {
+    operationProof?: OperationProof
+    operation?: Operation
+  }) {
     const currOperationProof = opts?.operationProof || this.operationProof
 
-    const identityParams = await getIdentityParams(querier)
+    const currOperation = opts?.operation || this.operation
 
-    const identityNode = await getIdentityNode(
-      querier,
-      identityParams.params.treapRootKey,
-    )
-
-    const newIdentitiesStatesRoot = identityNode.node.hash
+    const newIdentitiesStatesRoot = currOperation?.details?.stateRootHash
     const gistData = {
-      root: identityParams.params.GISTHash,
-      createdAtTimestamp: identityParams.params.GISTUpdatedTimestamp,
+      root: currOperation?.details?.GISTHash,
+      createdAtTimestamp: currOperation?.details?.timestamp,
     }
 
     const decodedPath = currOperationProof?.path?.map((el: string) =>