+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.0.0-rc.20] - 2023-08-25
+### Added
+- `@rarimo/client` - `getOperation` method and followed types
+- `@rarimo/zkp-gen-iden3` - `getOperation` method
+-
+### Changed
+- `@rarimo/zkp-gen-iden3` - getting params for transiting state (`loadParamsForTransitState`)
+
+### Removed
+- `@rarimo/client` - `getIdentityParams` method and followed types
+- `@rarimo/shared-zkp-iden3` - IdentityNode query
+
 ## [2.0.0-rc.19] - 2023-08-22
 ### Changed
 - `@rarimo/auth-zkp-iden3` - config options
@@ -432,7 +444,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Under the hood changes
 - Initiated repo with `@rarimo/provider` and `@rarimo/nft-checkout` packages
 
-[Unreleased]: https://github.com/rarimo/js-sdk/compare/2.0.0-rc.19...HEAD
+[Unreleased]: https://github.com/rarimo/js-sdk/compare/2.0.0-rc.20...HEAD
+[2.0.0-rc.20]: https://github.com/rarimo/js-sdk/compare/2.0.0-rc.19...2.0.0-rc.20
 [2.0.0-rc.19]: https://github.com/rarimo/js-sdk/compare/2.0.0-rc.18...2.0.0-rc.19
 [2.0.0-rc.18]: https://github.com/rarimo/js-sdk/compare/2.0.0-rc.17...2.0.0-rc.18
 [2.0.0-rc.17]: https://github.com/rarimo/js-sdk/compare/2.0.0-rc.16...2.0.0-rc.17